requirements.txt
psycopg2-binary
faker>=18.9.0
mysql-connector-python

